requirements.txt
django
djangorestframework
django-rest-swagger
drf-yasg
python-dotenv